增加出入院,24小时动态,身高体重,html页面缓存600秒

+++ custom_service_system/configs/lua_scripts/ac_del.lua
@@ -34,26 +34,60 @@ local function ParamCheck(post)
 end
 
 local function Execute(post)
+	-- local _id = post.web.id
+	-- local _loginid = post.session.loginid
+	
+	-- local _query_sql = "delete from user where id = "
+
+	-- for k, v in pairs(_id) do
+	-- 	if k == 1 then
+	-- 		if tonumber(v) == tonumber(_loginid) then
+	-- 			return nil, WEBERR.DEL_USER_SELF
+	-- 		end
+
+	-- 		_query_sql = _query_sql .. ngx.quote_sql_str(v)
+	-- 	else
+	-- 		_query_sql = _query_sql .. " or id = " .. ngx.quote_sql_str(v)
+	-- 	end
+	-- end
+
+	-- DEBUG("ac_del: ".._query_sql)
+	-- INFO(post.session.name.." 删除用户 ".._query_sql)
+	-- return mysql.query(cloud_database, _query_sql, MysqlCallback)
+
 	local _id = post.web.id
 	local _loginid = post.session.loginid
-	
-	local _query_sql = "delete from user where id = "
+
+	local function QueryCountBack(res)
+		if res[1].count ~= '0' then
+			return nil, WEBERR.DONT_DELETE_CUSTOMER
+		else
+			local _query_sql = "delete from user where id = "
+			for k, v in pairs(_id) do
+				if k == 1 then				
+					_query_sql = _query_sql .. ngx.quote_sql_str(v)
+				else
+					_query_sql = _query_sql .. " or id = " .. ngx.quote_sql_str(v)
+				end
+			end
+			DEBUG("ac_del: " .. _query_sql)
+			INFO(post.session.name.." 删除用户 ".._query_sql)
+			return mysql.query(cloud_database, _query_sql, MysqlCallback)
+		end
+	end
+
+	local _query_sql = "select count(*) As count from record where user_id = "
 
 	for k, v in pairs(_id) do
 		if k == 1 then
-			if tonumber(v) == tonumber(_loginid) then
-				return nil, WEBERR.DEL_USER_SELF
-			end
-
 			_query_sql = _query_sql .. ngx.quote_sql_str(v)
 		else
 			_query_sql = _query_sql .. " or id = " .. ngx.quote_sql_str(v)
 		end
 	end
 
-	DEBUG("ac_del: ".._query_sql)
-	INFO(post.session.name.." 删除用户 ".._query_sql)
-	return mysql.query(cloud_database, _query_sql, MysqlCallback)
+	DEBUG("QueryCount: " .. _query_sql)
+	return mysql.query(cloud_database, _query_sql, QueryCountBack)
 end
 local _M = {
 	ParamCheck = ParamCheck,
